Use blocks in tests/main.cpp to free the previous test before creating a new one
We don't want to have multiple connections with the exact same name around, as
that could possibly screw up an unrelated test. As there is no disconnect
method in the client class yet, we have to free the testobject that contains
the client object. the blocks in tests/main.cpp acheive just that.